update: 更新了type類型
