+++ lib/index.d.ts
@@ -1,0 +1,6 @@
+
+export {transformScript}  from './types/core'
+export {findSvgFile}  from './types/svg-file-loader'
+export {createFile}  from './types/create-file'
+export *  from './types/types'
+
