+++ lib/index.ts
@@ -7,4 +7,3 @@
 export { findSvgFile } from "./utils/svg-file-loader";
 export { createFile } from "./utils/create-file";
 export { transformScript } from "./utils/core";
-
