+++ lib/types/core.d.ts
@@ -1,0 +1,9 @@
+/*
+* @core.d.ts.ts
+* @deprecated 
+* @author czh
+* @update (czh 2022/3/23)
+*/
+import {Plugin} from "vite";
+import {ISvgTransScriptOption} from "./types";
+export function transformScript(options: ISvgTransScriptOption): Plugin
