+++ lib/types/create-file.d.ts
@@ -1,0 +1,8 @@
+/*
+* @create-file.d.ts.ts
+* @deprecated 
+* @author czh
+* @update (czh 2022/3/23)
+*/
+import {ISvgCtxItem} from "./types";
+export function createFile(outputPath:string,fileName:string ,type:string,format:string,svgCtx:Array<ISvgCtxItem>):void
