+++ lib/types/svg-file-loader.d.ts
@@ -1,0 +1,9 @@
+/*
+* @svg-file-loader.d.ts.ts
+* @deprecated 
+* @author czh
+* @update (czh 2022/3/23)
+*/
+
+import {ISvgCtxItem} from "./types";
+export function findSvgFile (dir:string):Array<ISvgCtxItem>
