+++ lib/types/types.d.ts
@@ -1,0 +1,17 @@
+/*
+* @types.d.ts.ts
+* @deprecated 
+* @author czh
+* @update (czh 2022/3/23)
+*/
+export declare interface ISvgTransScriptOption {
+    input: string
+    output?: string
+    name?: string
+    type?: string
+    format?: string
+}
+export declare interface ISvgCtxItem {
+    name: string
+    svg: string
+}
