+++ lib/utils/core.ts
@@ -1,7 +1,7 @@
 import {Plugin} from 'vite'
 import {findSvgFile} from "./svg-file-loader";
 import {createFile} from "./create-file";
-import {ISvgTransScriptOption,ISvgCtxItem} from "../types";
+import {ISvgTransScriptOption,ISvgCtxItem} from "../types/types";
 
 export const transformScript = ( config: ISvgTransScriptOption): Plugin => {
 // 读取svg文件内容
