+++ tsconfig.json
@@ -19,7 +19,8 @@
     "types": [
       "vite/client",
       "jest",
-      "/public/**/*.d.ts"
+      "/public/**/*.d.ts",
+      "/lib/types"
     ]
   },
   "include": [
